+++ examples/FirestoreDatabase/Documents/Async/NoCallback/Delete/Delete.ino
@@ -81,10 +81,6 @@ Firestore::Documents Docs;
 
 AsyncResult aResult_no_callback;
 
-int counter = 0;
-
-unsigned long dataMillis = 0;
-
 bool taskCompleted = false;
 
 void setup()
@@ -131,27 +127,22 @@ void loop()
 
     Docs.loop();
 
-    if (app.ready() && (millis() - dataMillis > 60000 || dataMillis == 0))
+    if (app.ready() && !taskCompleted)
     {
-        dataMillis = millis();
+        taskCompleted = true;
 
-        if (!taskCompleted)
-        {
-            taskCompleted = true;
+        // collection id > document id.
+        String documentPath = "test_doc_deletion/my_doc";
 
-            // aa is the collection id, bb is the document id in collection aa.
-            String documentPath = "aa/bb";
+        Serial.println("Creating a document... ");
 
-            Serial.println("Create a document... ");
+        Document<Values::Value> doc("myDouble", Values::Value(Values::DoubleValue(123.456)));
 
-            Document<Values::Value> doc("myDouble", Values::Value(Values::DoubleValue(123.456)));
+        Docs.createDocument(aClient, Firestore::Parent(FIREBASE_PROJECT_ID), documentPath, DocumentMask(), doc, aResult_no_callback);
 
-            Docs.createDocument(aClient, Firestore::Parent(FIREBASE_PROJECT_ID), documentPath, DocumentMask(), doc, aResult_no_callback);
+        Serial.println("Deleting a document... ");
 
-            Serial.println("Delete a document... ");
-
-            Docs.deleteDoc(aClient, Firestore::Parent(FIREBASE_PROJECT_ID), documentPath, Precondition() /* Precondition (currentocument) */, aResult_no_callback);
-        }
+        Docs.deleteDoc(aClient, Firestore::Parent(FIREBASE_PROJECT_ID), documentPath, Precondition() /* Precondition (currentocument) */, aResult_no_callback);
     }
 
     printResult(aResult_no_callback);
